Incluido app flask para disponibilizar REST APIspara manioulação de tabelas de pessoas e atividade. Tambem incluio o requirements do projeto

+++ utils.py
@@ -1,27 +1,34 @@
+from app import app
 from models import Pessoas
 
 def insere_pessoas():
-    pessoa = Pessoas(nome='Bruno', idade=61)
-    print(pessoa) 
+    pessoa = Pessoas(nome='Leo', idade=23)
+    print(pessoa)
     pessoa.save()
 
 def consulta_pessoas():
     pessoas = Pessoas.query.all()
     print(pessoas)
     pessoa = Pessoas.query.filter_by(nome='Renan').first()
-    print(pessoa.idade)
+    if pessoa:
+        print(pessoa.idade)
+    else:
+        print('Renan não encontrado')
 
 def altera_pessoa():
     pessoa = Pessoas.query.filter_by(nome='Alef').first()
-    pessoa.nome = 'Davi'
-    pessoa.save()
+    if pessoa:
+        pessoa.nome = 'Davi'
+        pessoa.save()
 
 def exclui_pessoa():
     pessoa = Pessoas.query.filter_by(nome='Davi').first()
-    pessoa.delete()
+    if pessoa:
+        pessoa.delete()
 
 if __name__ == '__main__':
-    #insere_pessoas()
-   # altera_pessoa()
-    exclui_pessoa()
-    consulta_pessoas()
+    with app.app_context():
+        #insere_pessoas()
+        # altera_pessoa()
+        # exclui_pessoa()
+        consulta_pessoas()
